rename package of resource locator
