+++ quarkus-extension-deployment/src/main/java/io/javaoperatorsdk/quarkus/resources/ResourceURLLocator.java
@@ -1,4 +1,4 @@
-package x.y.z;
+package io.javaoperatorsdk.quarkus.resources;
 
 import io.quarkus.runtime.annotations.RecordableConstructor;
 import java.net.URL;
