+++ quarkus-extension-deployment/src/test/java/io/javaoperatorsdk/quarkus/resources/ResourceProcessorTest.java
